requirements.txt
aiohttp
aiosqlite
discord.py
transformers
pillow
requests